move files to utils, generalize impose precision method
