+++ src/summarizedexperiment/utils/_slicer.py
