+++ src/summarizedexperiment/utils/_type_checks.py
