+++ src/summarizedexperiment/utils/_types.py
@@ -11,6 +11,7 @@ __copyright__ = "jkanche"
 __license__ = "MIT"
 
 MatrixTypes = Union[np.ndarray, sp.spmatrix, H5BackedSparseData, H5BackedDenseData]
+ArrayTypes = Union[np.ndarray, sp.lil_matrix]
 BiocOrPandasFrame = Union[pd.DataFrame, BiocFrame]
 MatrixSlicerTypes = Union[Sequence[int], Sequence[bool], slice]
 SlicerTypes = Union[Sequence[int], Sequence[bool], Sequence[str], slice]
